Clean up unused AgentEditor and AgentManager services
This moves a pause mediator callable from AgentManager to ExperimentManager.
