+++ frontend/src/components/chat/chat_info_panel.ts
@@ -10,7 +10,7 @@ import {CSSResultGroup, html, nothing} from 'lit';
 import {customElement, property, state} from 'lit/decorators.js';
 
 import {core} from '../../core/core';
-import {AgentManager} from '../../services/agent.manager';
+import {ExperimentManager} from '../../services/experiment.manager';
 import {AuthService} from '../../services/auth.service';
 import {CohortService} from '../../services/cohort.service';
 import {ParticipantService} from '../../services/participant.service';
@@ -34,7 +34,7 @@ import {styles} from './chat_info_panel.scss';
 export class ChatPanel extends MobxLitElement {
   static override styles: CSSResultGroup = [styles];
 
-  private readonly agentManager = core.getService(AgentManager);
+  private readonly experimentManager = core.getService(ExperimentManager);
   private readonly authService = core.getService(AuthService);
   private readonly cohortService = core.getService(CohortService);
   private readonly participantService = core.getService(ParticipantService);
@@ -140,7 +140,7 @@ export class ChatPanel extends MobxLitElement {
 
     const toggleStatus = async () => {
       this.isStatusLoading = true;
-      await this.agentManager.updateMediatorStatus(
+      await this.experimentManager.updateMediatorStatus(
         profile.id,
         profile.currentStatus === MediatorStatus.ACTIVE
           ? MediatorStatus.PAUSED
