+++ frontend/src/components/experiment_builder/agent_chat_prompt_editor.scss
