+++ frontend/src/service_provider.ts
@@ -7,8 +7,6 @@ import {ExperimentService} from './services/experiment.service';
 import {FirebaseService} from './services/firebase.service';
 import {HomeService} from './services/home.service';
 import {InitializationService} from './services/initialization.service';
-import {AgentEditor} from './services/agent.editor';
-import {AgentManager} from './services/agent.manager';
 import {ParticipantService} from './services/participant.service';
 import {ParticipantAnswerService} from './services/participant.answer';
 import {RouterService} from './services/router.service';
@@ -45,12 +43,6 @@ export function makeServiceProvider(self: Core) {
     },
     get initializationService() {
       return self.getService(InitializationService);
-    },
-    get agentEditor() {
-      return self.getService(AgentEditor);
-    },
-    get agentManager() {
-      return self.getService(AgentManager);
     },
     get participantService() {
       return self.getService(ParticipantService);
